Summary: WLT PoC works (but object repositinoing still have to be done) / surfaces manager seems to work and integrated with the "real" object detection
WLT: Updated OpenXR plugin to latest version - the use of a old version was the reason for the WLT for not working
StoringObject: integration of the Surfaces manager with the "real" detected objects.
Surfaces manager: to get it working, the box collider of each surface is on purpose bigger than the surface in y. Registration can be done using gameobjects with a callback called when the object crosses a surface; or by manually calling a function (which is the case with the "real" object).